server/smartsiteapi/requirements.txt
django
djangorestframework
pyrebase4
urllib3
requests==2.11.1
setuptools
pyyaml
PyJWT
google-cloud-firestore
google-auth
firebase-admin
google-cloud-storage